+++ src/callbacks/postPickupUpdate.ts
@@ -1,5 +1,6 @@
 import { ModCallback, PickupVariant } from "isaac-typescript-definitions";
 import * as removeGlitchedItems from "../features/mandatory/removeGlitchedItems";
+import * as taintedIsaacCollectibleDelay from "../features/optional/bugfix/taintedIsaacCollectibleDelay";
 import * as freeDevilItem from "../features/optional/major/freeDevilItem";
 
 export function init(mod: Mod): void {
@@ -14,6 +15,12 @@ export function init(mod: Mod): void {
 function collectibleCallback(pickup: EntityPickup) {
   const collectible = pickup as EntityPickupCollectible;
 
+  // Mandatory
   removeGlitchedItems.postPickupUpdateCollectible(collectible);
+
+  // Major
   freeDevilItem.postPickupUpdateCollectible(collectible);
+
+  // Bug fixes
+  taintedIsaacCollectibleDelay.postPickupUpdateCollectible(collectible);
 }