+++ src/classes/Config.ts
@@ -84,6 +84,7 @@ export class Config {
 
   // Bug fixes
   keeperHeal = true;
+  taintedIsaacCollectibleDelay = true;
   battery9VoltSynergy = true;
   reverseJusticeFix = true;
   preventUltraSecretRoomSoftlock = true;