+++ src/initFeatureVariables.ts
@@ -32,6 +32,7 @@ import * as killExtraEnemies from "./features/optional/bosses/killExtraEnemies";
 import * as preventVictoryLapPopup from "./features/optional/bosses/preventVictoryLapPopup";
 import * as battery9VoltSynergy from "./features/optional/bugfix/battery9VoltSynergy";
 import * as keeperHeal from "./features/optional/bugfix/keeperHeal";
+import * as taintedIsaacCollectibleDelay from "./features/optional/bugfix/taintedIsaacCollectibleDelay";
 import * as showEdenStartingItems from "./features/optional/characters/showEdenStartingItems";
 import * as globinSoftlock from "./features/optional/enemies/globinSoftlock";
 import * as extraStartingItems from "./features/optional/gameplay/extraStartingItems/extraStartingItems";
@@ -147,6 +148,7 @@ export function initFeatureVariables(): void {
 
   // Bug fixes
   keeperHeal.init();
+  taintedIsaacCollectibleDelay.init();
   battery9VoltSynergy.init();
 
   // Graphics fixes